+++ src/plugin.cpp
@@ -1,5 +1,5 @@
 #include "plugin.hpp"
-
+#include "drivers/MidiLoopback.hpp"
 
 Plugin* pluginInstance;
 
@@ -55,6 +55,10 @@ void init(rack::Plugin* p) {
 	p->addModel(modelMb);
 
 	pluginSettings.readFromJson();
+
+	if (pluginSettings.midiLoopbackDriverEnabled) {
+		StoermelderPackOne::MidiLoopback::init();
+	}
 }