+++ src/pluginsettings.cpp
@@ -14,6 +14,8 @@ void StoermelderSettings::saveToJson() {
     json_object_set(settingsJ, "mbV1sort", json_integer(mbV1sort));
     json_object_set(settingsJ, "mbV1hideBrands", json_boolean(mbV1hideBrands));
 
+    json_object_set(settingsJ, "midiLoopbackDriverEnabled", json_boolean(midiLoopbackDriverEnabled));
+
     std::string settingsFilename = rack::asset::user("Stoermelder-P1.json");
     FILE* file = fopen(settingsFilename.c_str(), "w");
     if (file) {
@@ -51,6 +53,9 @@ void StoermelderSettings::readFromJson() {
     if (mbV1sortJ) mbV1sort = json_integer_value(mbV1sortJ);
     json_t* mbV1hideBrandsJ = json_object_get(settingsJ, "mbV1hideBrands");
     if (mbV1hideBrandsJ) mbV1hideBrands = json_boolean_value(mbV1hideBrandsJ);
+
+    json_t* midiLoopbackDriverEnabledJ = json_object_get(settingsJ, "midiLoopbackDriverEnabled");
+    if (midiLoopbackDriverEnabledJ) midiLoopbackDriverEnabled = json_boolean_value(midiLoopbackDriverEnabledJ);
 
     fclose(file);
     json_decref(settingsJ);